restructure setting-related componnents
