+++ assets/app/js/components/Setting/RepoList.jsx
@@ -1,31 +1,11 @@
 import React, { Component } from 'react'
-import FlatButton from 'material-ui/FlatButton';
-
-class ListItem extends Component {
-    render() {
-        return (
-            <li>
-                <FlatButton
-                    className="select-repo-btn"
-                    label="Select"
-                    secondary={true}
-                    onTouchTap={this.props.syncWithGithub.bind(this,
-                    this.props.projectId,
-                    this.props.name,
-                    this.props.owner)}
-                />
-                <h4><a href={this.props.repoLink} target="_blank">{this.props.primaryText}</a></h4>
-                <p>{this.props.secondaryText}</p>
-            </li>
-        )
-    }
-}
+import GithubListItem from './GithubListItem.jsx';
 
 class List extends Component {
     render() {
         let repos = this.props.repos.map(repo => {
             return (
-                <ListItem
+                <GithubListItem
                     key={repo.id}
                     name={repo.name}
                     owner={repo.owner.login}
