+++ assets/app/js/components/Setting/Settings.jsx
@@ -6,11 +6,11 @@ import {Card, CardHeader, CardText} from 'material-ui/Card';
 import FontIcon from 'material-ui/FontIcon';
 import RaisedButton from 'material-ui/RaisedButton'
 
-import {getGithubAuthCode} from '../utils/general'
-import {githubOAuth} from '../utils/apiUtil'
-import {APP_ROOT_URL, PATH, GITHUB_CLIENT_ID, INVITED_TO_PROJECT, USER_ALREADY_EXISTS, USER_NOT_FOUND} from '../AppConstants';
-import LoadingIndicator from './Common/LoadingIndicator.jsx'
-import Github from './Github/Github.jsx'
+import {getGithubAuthCode} from '../../utils/general'
+import {githubOAuth} from '../../utils/apiUtil'
+import {APP_ROOT_URL, PATH, GITHUB_CLIENT_ID, INVITED_TO_PROJECT, USER_ALREADY_EXISTS, USER_NOT_FOUND} from '../../AppConstants';
+import LoadingIndicator from '../Common/LoadingIndicator.jsx'
+import Github from './Github.jsx'
 
 
 class Settings extends Component {
