+++ assets/app/js/containers/ProjectSettingView.jsx
@@ -1,7 +1,7 @@
 import React, { Component, PropTypes } from 'react'
 import { connect } from 'react-redux'
 
-import Settings from '../components/Settings.jsx'
+import Settings from '../components/Setting/Settings.jsx'
 
 const mapStateToProps = (state, ownProps) => {
   return {
